+++ src/main/core/errors.py
@@ -4,7 +4,7 @@ License, v. 2.0. If a copy of the MPL was not distributed with this
 file, You can obtain one at http://mozilla.org/MPL/2.0/.
 """
 
-from typing import Any
+from __future__ import annotations
 
 import discord
 from discord.ext.commands.errors import CommandError