+++ src/main/exts/meta/subcogs/meta__custom_commands.py
@@ -15,11 +15,11 @@ from discord.ext import commands
 
 from ....core import checks, db
 from ....core.context import Context
+from ....core.data import CacheListProperty, CacheUniqueViolation
 from ....core.embed import ZEmbed
 from ....core.guild import CCMode, GuildWrapper
 from ....core.menus import ZChoices, choice
 from ....core.mixin import CogMixin
-from ....utils.cache import CacheListProperty, CacheUniqueViolation
 from ....utils.format import CMDName, formatCmdName
 from ....utils.other import utcnow
 from .._custom_command import CustomCommand