+++ .github/CHANGELOG.md
@@ -24,6 +24,7 @@
 - [**Changed**] Merge TagScript into Z3R0 repository
 - [**Improved**] Split `_model` into `_wrapper` and `_custom_command`
 - [**Improved**] Moved Custom Command related errors into `exts.meta._errors`
+- [**Improved**] Moved JSON (and Blacklist) and Cache (and its properties) into `core.data`
 
 # v3.5 (Overhaul an Overhaul?)